Add Scrollview to UI
