+++ Assets/_Data/UI/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63104ad3f275d254eb63663418d01720
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
