+++ Assets/_Data/UI/Inventory/InventoryUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 71ef25862601f9747948df32cae60b07
