+++ Assets/_Data/UI/Text/TextAbstract.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cde8cb399f62eb64c8a852f4a56734d4
